// frontend/src/api/models/RegistrationRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Task Tracker API
 * Task Tracker API
 *
 * The version of the OpenAPI document: 1.0
 * Contact: trnka.roman13@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface RegistrationRequest
 */
export interface RegistrationRequest {
    /**
     * 
     * @type {string}
     * @memberof RegistrationRequest
     */
    firstname: string;
    /**
     * 
     * @type {string}
     * @memberof RegistrationRequest
     */
    lastname: string;
    /**
     * 
     * @type {string}
     * @memberof RegistrationRequest
     */
    email: string;
    /**
     * 
     * @type {string}
     * @memberof RegistrationRequest
     */
    password: string;
}

/**
 * Check if a given object implements the RegistrationRequest interface.
 */
export function instanceOfRegistrationRequest(value: object): value is RegistrationRequest {
    if (!('firstname' in value) || value['firstname'] === undefined) return false;
    if (!('lastname' in value) || value['lastname'] === undefined) return false;
    if (!('email' in value) || value['email'] === undefined) return false;
    if (!('password' in value) || value['password'] === undefined) return false;
    return true;
}

export function RegistrationRequestFromJSON(json: any): RegistrationRequest {
    return RegistrationRequestFromJSONTyped(json, false);
}

export function RegistrationRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): RegistrationRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'firstname': json['firstname'],
        'lastname': json['lastname'],
        'email': json['email'],
        'password': json['password'],
    };
}

export function RegistrationRequestToJSON(value?: RegistrationRequest | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'firstname': value['firstname'],
        'lastname': value['lastname'],
        'email': value['email'],
        'password': value['password'],
    };
}